testing: clean up devstack testing
- less constructors
