+++ pkg/test/devstack/submit_test.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/bacalhau-project/bacalhau/pkg/node"
+	"github.com/bacalhau-project/bacalhau/pkg/devstack"
 	testutils "github.com/bacalhau-project/bacalhau/pkg/test/utils"
 
 	"github.com/stretchr/testify/require"
@@ -38,12 +38,8 @@ func (suite *DevstackSubmitSuite) SetupTest() {
 func (suite *DevstackSubmitSuite) TestEmptySpec() {
 	ctx := context.Background()
 
-	stack, _ := testutils.SetupTest(
-		ctx,
-		suite.T(),
-		1,
-		node.NewComputeConfigWithDefaults(),
-		node.NewRequesterConfigWithDefaults(),
+	stack := testutils.SetupTestDevStack(ctx, suite.T(),
+		devstack.WithNumberOfHybridNodes(1),
 	)
 
 	apiServer := stack.Nodes[0].APIServer
