+++ pkg/test/scenario/suite.go
@@ -5,6 +5,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/stretchr/testify/suite"
+
 	"github.com/bacalhau-project/bacalhau/pkg/devstack"
 	"github.com/bacalhau-project/bacalhau/pkg/docker"
 	"github.com/bacalhau-project/bacalhau/pkg/downloader"
@@ -16,7 +18,6 @@ import (
 	"github.com/bacalhau-project/bacalhau/pkg/system"
 	"github.com/bacalhau-project/bacalhau/pkg/telemetry"
 	testutils "github.com/bacalhau-project/bacalhau/pkg/test/utils"
-	"github.com/stretchr/testify/suite"
 )
 
 type ScenarioTestSuite interface {
@@ -82,14 +83,17 @@ func (s *ScenarioRunner) setupStack(config *StackConfig) (*devstack.DevStack, *s
 	if config.ComputeConfig.TotalResourceLimits == empty {
 		config.ComputeConfig = node.NewComputeConfigWithDefaults()
 	}
-
-	stack := testutils.SetupTestWithNoopExecutor(
-		s.Ctx,
-		s.T(),
-		*config.DevStackOptions,
-		config.ComputeConfig,
-		config.RequesterConfig,
-		config.ExecutorConfig,
+	stack := testutils.SetupTestDevStack(s.Ctx, s.T(),
+		append(config.DevStackOptions.Options(),
+			devstack.WithComputeConfig(config.ComputeConfig),
+			devstack.WithRequesterConfig(config.RequesterConfig),
+			devstack.WithDependencyInjector(node.NodeDependencyInjector{
+				ExecutorsFactory: &testutils.MixedExecutorFactory{
+					StandardFactory: node.NewStandardExecutorsFactory(),
+					NoopFactory:     devstack.NewNoopExecutorsFactoryWithConfig(config.ExecutorConfig),
+				},
+			}),
+		)...,
 	)
 
 	return stack, stack.Nodes[0].CleanupManager
